+++ EndToEndTests/SampleTextTest.php
@@ -10,7 +10,7 @@ class AnyMark_EndToEndTests_SampleTextTest extends \AnyMark\EndToEndTests\Suppor
 	public function acceptsCustomPatternsFile()
 	{
 		// given
-		$anyMark = \AnyMark\AnyMark::setup()->get('AnyMark\\AnyMark');
+		$anyMark = \AnyMark\AnyMark::createWith(\AnyMark\AnyMark::defaultSetup());
 		$anyMark->setPatternsFile(__DIR__
 			. DIRECTORY_SEPARATOR . '..'
 			. DIRECTORY_SEPARATOR . 'AnyMark'